refactor: Extract function into helpers module file
