+++ lib/index.js
@@ -20,6 +20,7 @@ const createHash = require('crypto').createHash;
 
 const getLicenseTitle = require('./getLicenseTitle');
 const licenseFiles = require('./license-files');
+const helpers = require('../lib/helpers');
 
 // Set up debug logging
 // https://www.npmjs.com/package/debug#stderr-vs-stdout
@@ -361,41 +362,6 @@ const recursivelyCollectAllDependencies = (options) => {
     return data;
 };
 
-/**
- * ! This function has a wanted sideeffect, as it modifies the json object that is passed by reference.
- *
- *  The depth attribute set in the options parameter here - which is defined by setting the `--direct` flag - is of
- *  no use with npm < 3, as the older npm versions flattened all dependencies into one single directory. So in
- *  order to making `--direct` work with older versions of npm, we need to filter out all non-dependencies from
- *  the json result.
- */
-// TODO: Add tests for this function
-const deleteNonDirectDependenciesFromAllDependencies = (
-    { _dependencies: directDependencies, dependencies: allDependencies, devDependencies },
-    options,
-) => {
-    const allDependenciesArray = Object.keys(allDependencies);
-    const directDependenciesArray = Object.keys(directDependencies);
-    const devDependenciesArray = Object.keys(devDependencies);
-    let wantedDependenciesArray = [];
-
-    if (options.production && !options.development) {
-        wantedDependenciesArray = directDependenciesArray.filter(
-            (directDependency) => !devDependenciesArray.includes(directDependency),
-        );
-    } else if (!options.production && options.development) {
-        wantedDependenciesArray = devDependenciesArray;
-    } else {
-        wantedDependenciesArray = directDependenciesArray;
-    }
-
-    allDependenciesArray.forEach((currentDependency) => {
-        if (!wantedDependenciesArray.includes(currentDependency)) {
-            delete allDependencies[currentDependency];
-        }
-    });
-};
-
 exports.init = function init(args, callback) {
     // Fix path if on Windows:
     const workingDir = args.start.replace(/\\\\/g, '\\');
@@ -455,7 +421,7 @@ exports.init = function init(args, callback) {
         // the package.json file in the property '_dependencies'. The property 'dependencies' contains all dependencies,
         // including the ones that are only required by other dependencies.
         if (optionsForReadingInstalledPackages.depth === 0) {
-            deleteNonDirectDependenciesFromAllDependencies(installedPackagesJson, args);
+            helpers.deleteNonDirectDependenciesFromAllDependencies(installedPackagesJson, args);
         }
 
         const data = recursivelyCollectAllDependencies({
